stkpro/requirements.txt
aioredis==1.3.1
amqp==5.1.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
autobahns
Automat==20.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.0
charset-normalizer==2.0.12
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
constantly==15.1.0
cryptography
cycler==0.11.0
daphne==3.0.2
dateutils==0.6.12
Deprecated==1.2.13
Django==3.2.12
django-celery-beat==2.2.1
django-celery-results==2.3.0
django-filter==21.1
django-redis
django-timezone-field
djangorestframework==3.13.1
fonttools
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
importlib-metadata
incremental==21.3.0
kiwisolver
kombu==5.2.4
lxml
Markdown==3.3.6
matplotlib==3.5.1
msgpack==1.0.3
nsepy==0.8
nsetools==1.0.11
numpy
packaging==21.3
pandas
Pillow
prompt-toolkit==3.0.28
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.7
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
redis==4.1.4
requests==2.27.1
service-identity==21.1.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
Twisted==22.2.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.1.1
tzdata
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.0
zipp==3.7.0
zope.interface==5.4.0
